brain-mri-deployment/requirements.txt
fastapi
uvicorn
torch
numpy
opencv-python-headless
pillow
python-multipart